+++ src/main/java/com/juyo/visa/MainModule.java
@@ -2,9 +2,11 @@ package com.juyo.visa;
 
 import org.nutz.ioc.loader.annotation.IocBean;
 import org.nutz.mvc.annotation.At;
+import org.nutz.mvc.annotation.By;
 import org.nutz.mvc.annotation.ChainBy;
 import org.nutz.mvc.annotation.Encoding;
 import org.nutz.mvc.annotation.Fail;
+import org.nutz.mvc.annotation.Filters;
 import org.nutz.mvc.annotation.IocBy;
 import org.nutz.mvc.annotation.Localization;
 import org.nutz.mvc.annotation.Modules;
@@ -13,6 +15,8 @@ import org.nutz.mvc.annotation.SetupBy;
 import org.nutz.mvc.annotation.Views;
 import org.nutz.mvc.ioc.provider.ComboIocProvider;
 
+import com.juyo.visa.common.actionfilter.AuthFilter;
+import com.juyo.visa.common.actionfilter.LoginFilter;
 import com.uxuexi.core.web.view.WeViewMaker;
 
 @IocBy(type = ComboIocProvider.class, args = { "*org.nutz.ioc.loader.json.JsonLoader", "webconfig/",
@@ -25,7 +29,7 @@ import com.uxuexi.core.web.view.WeViewMaker;
 @SetupBy(WeSetup.class)
 @ChainBy(args = { "${app.root}/WEB-INF/classes/webconfig/chains.js" })
 @Views(WeViewMaker.class)
-//@Filters({ @By(type = LoginFilter.class), @By(type = AuthFilter.class) })
+@Filters({ @By(type = LoginFilter.class), @By(type = AuthFilter.class) })
 @IocBean
 public class MainModule {
