+++ src/main/java/com/juyo/visa/admin/company/module/CompanyModule.java
@@ -5,7 +5,6 @@ import org.nutz.ioc.loader.annotation.IocBean;
 import org.nutz.log.Log;
 import org.nutz.log.Logs;
 import org.nutz.mvc.annotation.At;
-import org.nutz.mvc.annotation.Filters;
 import org.nutz.mvc.annotation.GET;
 import org.nutz.mvc.annotation.Ok;
 import org.nutz.mvc.annotation.POST;
@@ -20,8 +19,6 @@ import com.uxuexi.core.web.chain.support.JsonResult;
 
 @IocBean
 @At("/admin/company")
-@Filters({//@By(type = AuthFilter.class)
-})
 public class CompanyModule {
 
 	private static final Log log = Logs.get();
