+++ src/main/webapp/WEB-INF/public/header.jsp
@@ -46,7 +46,7 @@
             <span class="dian"></span>
           </li>
           <li class="dropdown messages-menu">
-            <a href="#" class="dropdown-toggle name" data-toggle="dropdown">${loginuser.fullName}</a>
+            <a href="#" class="dropdown-toggle name" data-toggle="dropdown">${loginuser.name}</a>
           </li>
           <li class="setUp-li">
              <a href="javascript:;">设置<img class="setUp" src="${base}/references/public/dist/newvisacss/img/setUp.jpg"></a>
